requirements.txt
FastMCP
gradio
pathlib
mcp
google-generativeai
python-dotenv
pylint
langchain
langchain-google-genai
langchain-mcp-adapters
